Changed the small string library to work with larger lengths.

+++ src/smlstr.h
@@ -17,11 +17,11 @@
 
 START_C_SECTION
 
-uint8_t smlstrcpy(string_t dest, kstring_t src, uint8_t len);
-uint8_t smlstrcat(string_t dest, kstring_t src, uint8_t len);
+uint16_t smlstrcpy(string_t dest, kstring_t src, uint16_t len);
+uint16_t smlstrcat(string_t dest, kstring_t src, uint16_t len);
 
-uint8_t smluintfmt(string_t dest, uint32_t val, uint8_t len);
-uint8_t smlintfmt(string_t dest, int32_t val, uint8_t len);
+uint16_t smluintfmt(string_t dest, uint32_t val, uint16_t len);
+uint16_t smlintfmt(string_t dest, int32_t val, uint16_t len);
 
 bool_t smlisdec(kstring_t src);
 uint32_t smluintscan(kstring_t src);
